services/backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.1
mysql-connector-python==8.1.0